# Encourage inlining of Iter()